project/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
cryptography==42.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.2.6
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.27.2
exceptiongroup==1.2.0
executing==2.0.1
idna==3.7
importlib-metadata==7.0.1
inflection==0.5.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.17.0
